% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circletab.R
\name{circletab}
\alias{circletab}
\title{Table with circles}
\usage{
circletab(
  title,
  description,
  data = data.frame(),
  max_value,
  fill_color,
  stroke_width,
  stroke_color,
  width = "100\%",
  height = "auto",
  elementId = NULL
)
}
\arguments{
\item{title}{Title}

\item{description}{Description}

\item{data}{Data frame with three columns with two categorical variables and a single measure (count, proportion, etc.)}

\item{width}{Width}

\item{height}{Height}

\item{elementId}{Block ID}

\item{max_val}{Lower and upper values for bar track}

\item{bar_color}{Bar color}

\item{track_color}{Track color}

\item{font_color}{Font color}
}
\description{
Table where values are represented as circles of varying size.
}
\examples{
"hi"
}
