% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telcowidgets.R
\name{net_promoter_score}
\alias{net_promoter_score}
\title{Net Promoter Score}
\usage{
net_promoter_score(
  title,
  description,
  data = data.frame(brand = c("contact", "mercury", "onenz", "skinny", "slingshot",
    "spark", "2degrees"), value = c(0, 0, 0, 0, 0, 0, 0)),
  unit,
  range,
  bar_color = rgb(63, 94, 88, maxColorValue = 255),
  track_color = rgb(221, 229, 235, maxColorValue = 255),
  font_color = bar_color,
  width = "100\%",
  height = "auto",
  elementId = NULL
)
}
\arguments{
\item{title}{Title}

\item{description}{Description}

\item{data}{Data frame with two columns: brand and value}

\item{unit}{Unit to append to bar values, e.g. "\%"}

\item{range}{Lower and upper values for bar track}

\item{bar_color}{Bar color}

\item{track_color}{Track color}

\item{font_color}{Font color}

\item{width}{Width}

\item{height}{Height}

\item{elementId}{Block ID}
}
\description{
Net promoter score compoenent for Telco providers.
}
\examples{
net_promoter_score(
  title = "Speed of resolution",
  description = "Measures satisfaction as how quickly providers resolve customer service issues (higher is better).",
  data = data.frame(
    brand = c("onenz", "skinny", "spark", "2degrees"),
    value = c(45, 49, 43, 54)
  ),
  unit = "\%",
  range = c(0,100),
  bar_color = rgb(176, 12, 41, maxColorValue = 255),
  font_color = rgb(176, 12, 41, maxColorValue = 255)
)

net_promoter_score(
  title = "Net Promoter Score (NPS)",
  description = "Mesaures how likely customers are to recommend their provider to friends and family (higher is better).",
  data = data.frame(
    brand = c("contact", "mercury", "onenz", "skinny", "slingshot", "spark", "2degrees"),
    value = c(-1, -11, -2, 11, 2, 2, 4)
  ),
  range = c(-20, 20),
  bar_color = rgb(63, 94, 88, maxColorValue = 255),
  font_color = rgb(63, 94, 88, maxColorValue = 255)
)
}
